Add improvements to custom schema
