+++ .gitignore
@@ -6,6 +6,7 @@
 *.iws
 *.ipr
 .idea
+.DS_Store
 
 # Package Files #
 *.jar
