+++ modules/charon-core/src/main/java/org/wso2/charon3/core/schema/SCIMConstants.java
@@ -702,6 +702,9 @@ public class SCIMConstants {
         public static final String ENTERPRISE_USER_DESC = "Enterprise User";
     }
 
+    /**
+     * Constants found in custom user schema.
+     */
     public static class CustomUserSchemaConstants {
 
         public static final String NAME = "name";
